Remove dead code from output files
I removed the "tab" format entirely, which represented ascii tables.
I also removed the old history/table logic such as "sum" and "slice."
We may want this functionality at some point, but the data model has
changed so much that I think we probably just want to start from scratch.

One piece of machinery that may need to be re-introduced, but probably
in a different place in the code base, is the transformation operation
from curvilinear coordinates to Cartesian. For vis, downstream codes
will need to know how to do that. But I don't think the "weakly linked"
function stub that was present here was the right place for it.

I left the VTK stub alone as we may want that and the old logic is
probably still a useful reference.