Let Parthenon tasks fail
Previously, due to using a worker thread, Parthenon would drop any
exceptions from Tasks executed in a TaskList.  Exceptions would
crash the thread, but execution would continue as if it had
finished normally.

This commit checks the futures associated with each task, to make
sure there are not exceptions.  As a bonus, we can add a
potentially non-fatal failure return TaskStatus::fail, which will
be propagated up to the Driver.