Fix some restarts w/non-CC fields
When restarting, Parthenon reads fields into a temporary space
allocated to be as large as the largest field.  However,
if the largest field was e.g. face-centered it would not be
calculated correctly, resulting in overflowing the temporary.

Additionally, I think this would have made sparse non-CC fields
read wrong.

I just simplify both with the new, better `VarInfo` utilities.